+++ atlas-browser/src/main/java/com/atlas/browser/BrowserAutomation.java
@@ -2,4 +2,6 @@ package com.atlas.browser;
 
 public interface BrowserAutomation {
     BrowserResult inspect(BrowserRequest request);
+
+    BrowserApplicationResult apply(BrowserApplicationRequest request);
 }